+++ tests/Twilio/Integration/Lookups/V2/PhoneNumberTest.php
@@ -485,7 +485,7 @@ class PhoneNumberTest extends HolodeckTestCase {
                     "address_lines_match": "no_match",
                     "city_match": "no_match",
                     "state_match": "high_partial_match",
-                    "postal_code_match": "no_data",
+                    "postal_code_match": "no_data_available",
                     "address_country_match": "exact_match",
                     "national_id_match": "exact_match",
                     "date_of_birth_match": "exact_match",