+++ CHANGES.md
@@ -1,6 +1,18 @@
 twilio-php Changelog
 ====================
 
+[2023-01-11] Version 6.44.1
+---------------------------
+**Conversations**
+- Add support for creating Multi-Channel Rich Content Messages
+
+**Lookups**
+- Changed the no data message for match postal code from `no_data` to `data_not_available` in identity match package
+
+**Messaging**
+- Add update/edit tollfree verification API
+
+
 [2022-12-14] Version 6.44.0
 ---------------------------
 **Api**